Initial draft of BuildTree class
